added evaluation of single expression + clippy pedantic compliance

+++ src/interpreter.rs
@@ -13,7 +13,7 @@ impl Interpreter {
         }
     }
 
-    pub(crate) fn interpret_ast(&mut self, ast: Expression) -> Result<i64, String> {
+    pub(crate) fn interpret_ast(&mut self, ast: Expression) -> Result<f64, String> {
        match ast {
            Assignment(identifier, expr) => {
                self.mem.entry(identifier).and_modify(|val| *val = *expr.clone()).or_insert(*expr.clone());
@@ -27,10 +27,10 @@ impl Interpreter {
 
            }
            Expression::UnaryPlus(expr) => {
-               Ok(0 + self.interpret_ast(*expr)?)  // Let's pretend it is somehow useful
+               Ok(0f64 + self.interpret_ast(*expr)?)  // Let's pretend it is somehow useful
            }
            Expression::UnaryMinus(expr) => {
-               Ok(0 - self.interpret_ast(*expr)?)
+               Ok(0f64 - self.interpret_ast(*expr)?)
            }
            Expression::ParenthesisExpression(expr) => {
                Ok(self.interpret_ast(*expr)?)
@@ -41,7 +41,7 @@ impl Interpreter {
            Expression::Division(left, right) => {
                Ok(self.interpret_ast(*left)? / self.interpret_ast(*right)?)
            }
-           Literal(value) => { Ok(value as i64) }
+           Literal(value) => { Ok(value) }
            Expression::Variable(identifier) => {
                if let Some(expr) = self.mem.get(identifier.as_str()) {
                    Ok(self.interpret_ast(expr.clone())?)
